% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_combination.R
\name{all_combination}
\alias{all_combination}
\title{Title Function to return all combinations of objects}
\usage{
all_combination(N, x, weights, items, value)
}
\arguments{
\item{N}{Number of rows in the input dataset}

\item{x}{The input dataset}

\item{weights}{The vector containing the weights of the items}

\item{items}{The vector containing the list of the items}

\item{value}{The vector containing the value of the items}
}
\description{
Title Function to return all combinations of objects
}
