% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Function to solve the Knapsack problem using the Greedy Heuristic Approach}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v(alue) and w(eight)}

\item{W}{the size of the napsack}
}
\value{
list with maximum value and the elements
}
\description{
Function to solve the Knapsack problem using the Greedy Heuristic Approach
}
\examples{
x <- create_problem(2000)
greedy_knapsack(x[1:800,],3500)
}
