% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{knapsack_brute_force}
\alias{knapsack_brute_force}
\title{Function to solve the Knapsack problem using Brute Force Method}
\usage{
knapsack_brute_force(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data.frame with two variables v(alue) and w(eight)}

\item{W}{the size of the napsack}

\item{parallel}{parallelization flag, default = FALSE}
}
\value{
list with maximum value and the elements
}
\description{
Function to solve the Knapsack problem using Brute Force Method
}
\examples{
x <- create_problem(2000)
print(head(x))
knapsack_brute_force(x = x[1:8,],W = 3500, parallel = FALSE)
}
