% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Function to solving Knapsack Problem using Dynammic Programming}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{a data.frame with two variables v(alue) and w(eight)}

\item{W}{the size of the napsack}
}
\value{
list with maximum value and the elements
}
\description{
Function to solving Knapsack Problem using Dynammic Programming
}
\examples{
x <- create_problem(2000)
knapsack_dynamic(x[1:8,],3500)
}
