% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_profiling.R
\name{profile_bruteforce}
\alias{profile_bruteforce}
\title{Profling function for Brute force method}
\usage{
profile_bruteforce(number = 16, W = 3500, parallel = FALSE)
}
\arguments{
\item{number}{number of row to select from randomly generated dataset}

\item{W}{maximum weight limit}

\item{parallel}{parallel flag, default = FALSE.}
}
\value{
profvis object
}
\description{
Profling function for Brute force method
}
\examples{
profile_bruteforce(16,3500)
}
