% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_profiling.R
\name{profile_dynamic}
\alias{profile_dynamic}
\title{Profling function for Dynamic method}
\usage{
profile_dynamic(number = 500, W = 3500)
}
\arguments{
\item{number}{number of row to select from randomly generated dataset}

\item{W}{maximum weight limit}
}
\value{
profvis object
}
\description{
Profling function for Dynamic method
}
\examples{
profile_dynamic(500,3500)
}
