% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_profiling.R
\name{profile_greedy}
\alias{profile_greedy}
\title{Profling function for Greedy Heuristic method}
\usage{
profile_greedy(number = 1e+06, W = 3500)
}
\arguments{
\item{number}{number of row to select from randomly generated dataset}

\item{W}{maximum weight limit}
}
\value{
profvis object
}
\description{
Profling function for Greedy Heuristic method
}
\examples{
profile_greedy(1000000,3500)
}
