% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Title The Brute Force Knapsnack Solver}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{The input to function containing the number of items}

\item{W}{The weights of the items}
}
\description{
Title The Brute Force Knapsnack Solver
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
